+++ flake.nix
@@ -167,6 +167,8 @@
           jackify
           minecraft-server-ctl
           minecraft-server-image
+          porkbun-domain-search-mcp
+          porkbun-mcp-domain-search
           qwen-code
           reference-download
           re-container